requirements.txt
django_allauth==0.45.0
Django==2.2.24
djangorestframework==3.12.4
psycopg2-binary
#psycopg2>=2.8,<2.9
django-rest-swagger
pyyaml
Pillow 